Changes to prep for episode 1 of KRC
Changes:

*Commented out the overlay and the header darkening. They're cool, but
not the right fit for my TV show.  Uncomment line 267 and 268  to
restore that functionality. Eventually, we can have a hotkey to enable
it.

*Made the entire project work locally since our teleprompter computer is
not Internet connected. Downloaded all supporting libraries and fonts.
Added license notices where necessary.

*Replaced cookie calls with localstorage since localstorage works
without a webserver. Entire project can now be run locally and it will
remember settings between refreshes.

*Made the keyboard work in a more reasonable way when it's playing. You
don't have to click in weird places to make the keyboard shortcuts work
anymore using a weird focus trick I discovered. I'm still not totally
sure what's up with that. Changed up and down to actually scroll up and
down. Font size is controlled by + and -. It's instinctual to want to
use up and down to *go* up and down, so I put in that behavior.

*Sped up the prompter to 2 pixels per tick so it has a faster max speed
on this crappy old laptop and changed the range so it can run slower as
well. Changed the speed interval to 10. Speed of zero now effectively
pauses the prompter.